Cart Service Partially Converted to Hystrix
Fails Integration Tests.
Zipkin Tracing works as intended.

Every function (except toString/toJson) has been converted to use the Hystrix lambda expression.  This will work as a template to begin converting the other services.  This commit has errors, it has been pushed to obtain help with debugging.

Errors: the placeOrder function always goes into the fallback for an unknown reason.  There are ample print statements inside the placeOrder function itself, as well as GenericHystrixCommand.java:90, which prints the errors that resulted in the Hystrix fallback.  The undoCart function follows a similar execution pattern to the placeOrder function, but does not have the same error.