Hystrix Integration for All Services
Integration Tests Pass.
8 Spans/4 depth Pass.

All microservices now have complete integration with Hystrix.  No "try again" model fallbacks have been implemented in this commit, the functionality is exactly the same as previous commits.  The next step will be to add backup services and Hystrix fallback commands for them.  This commit prepares for that by moving the functionality of each RequestMapping to a separate function that can be passed different IP addresses of both the primary and backup instances of the services.

NOTE: Hystrix still thinks that the regular startup time of the docker containers/instances constitutes a timeout.  This commit tries to account for that inside the GenericHystrixCommand by making the timeout 4000ms rather than the default 1000ms, but it hasn't had any effect.  Because each Hystrix module will do this, the example command will fail several times before succeeding normally.